Updated animation_loader
Now animations can have their own frame width, which has to
be specified either inside animations_setup.c or, hopefully
I'll get this data from the parser in the future,

meaning that I'll be able to load animation from a spread sheet
of an arbitrary width, with arbitrary amount of frames,
as long as you specify the width of one frame.

But it still has to be a one row spread sheet though.