Now the file paths specified inside the .cub file, are considered to be file paths to the animations. And if it is a static image, then it will simply be an animation with 1 frame.
And therefore I updated a bunch of code to handle that.

Also, now, to know how many frames there are, I simply use
sprite_sheet.height, because of our agreement that
the textures can only be squares,
and a sprite sheet can only contain 1 row.
Otherwise it will not work, unless I get this additional info
from the parser, meaning we would have to extend the .cub
syntax specification.